Presimulacion no optimizada hecha

+++ fluid/fluid.cpp
@@ -6,6 +6,7 @@
 #include "../sim/math.hpp"
 #include "../sim/grid.hpp"
 #include "../sim/files.hpp"
+
 int main(int argc, char* argv[]) {
   std::vector<std::string> argumentos;
   for (int i = 1; i < argc; i++) argumentos.push_back(argv[i]); ///Nos olvidamos del problema de punteros de argv pasandolo todo a un vector
@@ -15,7 +16,8 @@ int main(int argc, char* argv[]) {
   std::array<int,3> bloques;
   bloques = calc_n_blocks(datos_grid.ppm); ///Calculamos ppm de ese struct de datos de cabecera y almacenamos esos datos en el array bloques
   Grid grid(bloques[0],bloques[1],bloques[2]); ///Con el array bloques creamos el objeto 'grid'
-  std:: array<int, 3> medidas_bloque = calc_size_blocks(bloques);
+  std:: array<double, 3> medidas_bloque = calc_size_blocks(bloques);
+
   int id = Read_particles(argumentos[1], medidas_bloque, grid);
   std:: cout << id << std::endl;
   return 0;
